This is a way to handle mocking.
The essential idea is we want to be able to test code without
that code calling out to something we don't want to test.

One way to do that is what we've done here.  When we're testing
a method that, for example, calls to a database through an object,
for those methods we'll pass in the object that grants access to
the database, and thus we can pass in anything we want.

If you think about it, all code naturally calls out to other classes,
if even just the standard classes and tools, so realistically this
should be used on a prudent basis, not a pedantic one.